Many updates since this was out of date
Gave a simple example add of adding reminder for the logged in user.
updated the greplytix-application for client side since it was
out-of-date.

The example app deminstrates how to load data dynamically from the
server, and save a record to the server. It also gives a clear example
of dynamic rendering with a collection as new models are created.

The example is not complete but gives a good starting basis with best
practices approaches.